Add parser for Snapchat history HTML file
This commit introduces a new parser written in Go that reads a Snapchat history HTML file and extracts relevant information, which it then saves into a new CSV file. A new entry in the .gitignore file ensures that build files do not get committed. The Makefile has also been updated to include commands to build and run the parser, and clean up the build files.